# ui/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Binance Pay Gateway API",
    "description": "Payment gateway menggunakan akun Binance personal. Mendukung Binance Pay (Pay ID) dan on-chain deposit (TRC20, BEP20, dll).\n\n## Cara Pakai\nSet credentials di header setiap request:\n- `X-Binance-Api-Key` — API Key Binance kamu\n- `X-Binance-Api-Secret` — API Secret Binance kamu  \n- `X-Binance-Pay-Id` — Pay ID kamu (9 digit)\n\nAtau isi via form di halaman utama, credentials disimpan di localStorage.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "API Server" }
  ],
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Binance-Api-Key",
        "description": "Binance API Key (Enable Reading only)"
      },
      "ApiSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Binance-Api-Secret",
        "description": "Binance API Secret"
      },
      "PayId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Binance-Pay-Id",
        "description": "Binance Pay ID (9 digit)"
      }
    },
    "schemas": {
      "Invoice": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "INV1A2B3C4DABCD" },
          "status": { "type": "string", "enum": ["PENDING", "PAID", "EXPIRED"] },
          "currency": { "type": "string", "example": "USDT" },
          "baseAmount": { "type": "number", "example": 5.00 },
          "expectedAmount": { "type": "number", "example": 5.0037 },
          "network": { "type": "string", "example": "TRC20", "nullable": true },
          "productId": { "type": "string", "nullable": true },
          "buyer": { "type": "string", "nullable": true },
          "createdAt": { "type": "integer", "example": 1749563004000 },
          "expiresAt": { "type": "integer", "example": 1749564804000 },
          "transactionId": { "type": "string", "nullable": true },
          "paidAmount": { "type": "number", "nullable": true },
          "paidAt": { "type": "integer", "nullable": true },
          "payMethod": { "type": "string", "enum": ["PAY", "ONCHAIN"], "nullable": true },
          "txHash": { "type": "string", "nullable": true }
        }
      },
      "PaymentOption": {
        "type": "object",
        "properties": {
          "method": { "type": "string", "enum": ["BINANCE_PAY", "ONCHAIN"] },
          "description": { "type": "string" },
          "payId": { "type": "string", "nullable": true },
          "network": { "type": "string", "nullable": true },
          "depositAddress": { "type": "string", "nullable": true },
          "amountToPay": { "type": "number" },
          "currency": { "type": "string" },
          "instruction": { "type": "string" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "security": [
    { "ApiKey": [], "ApiSecret": [], "PayId": [] }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "tags": ["System"],
        "security": [],
        "responses": {
          "200": {
            "description": "Server berjalan normal",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "ts": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/payment-options": {
      "post": {
        "summary": "Buat instruksi pembayaran",
        "description": "Menghasilkan saran nominal unik + opsi pembayaran (Pay ID / deposit address). STATELESS — server tidak menyimpan apa pun. Sistem pemanggil yang menyimpan nominal untuk dicek nanti.",
        "tags": ["Payment"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": { "type": "number", "example": 5.00, "description": "Nominal dasar" },
                  "currency": { "type": "string", "example": "USDT", "description": "Mata uang (default USDT)" },
                  "network": { "type": "string", "example": "TRC20", "description": "Batasi ke satu network. Kosongkan untuk semua. PAY = Binance Pay only." }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instruksi pembayaran",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "amountToPay": { "type": "number", "description": "Nominal unik yang harus dibayar buyer" },
                    "currency": { "type": "string" },
                    "network": { "type": "string", "nullable": true },
                    "tolerancePercent": { "type": "number" },
                    "paymentOptions": { "type": "array", "items": { "$ref": "#/components/schemas/PaymentOption" } }
                  }
                }
              }
            }
          },
          "400": { "description": "Input tidak valid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/check-payment": {
      "post": {
        "summary": "Cek pembayaran masuk",
        "description": "Cek apakah ADA pembayaran masuk yang cocok dengan kriteria (amount, currency, network, window waktu). STATELESS — tidak menyimpan apa pun. Anti-replay (cegah 1 transaksi dipakai 2x) adalah tanggung jawab sistem pemanggil.",
        "tags": ["Payment"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": { "type": "number", "example": 5.0037, "description": "Nominal yang dicocokkan (gunakan amountToPay dari /payment-options)" },
                  "currency": { "type": "string", "example": "USDT" },
                  "network": { "type": "string", "example": "TRC20", "description": "Batasi metode. PAY = Binance Pay only. Kosong = semua." },
                  "sinceMinutes": { "type": "number", "example": 60, "description": "Window pencarian ke belakang (menit, max 10080 = 7 hari)" },
                  "tolerancePercent": { "type": "number", "example": 0.5, "description": "Override toleransi selisih nominal" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hasil pengecekan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "paid": { "type": "boolean", "description": "true kalau ada pembayaran cocok" },
                    "match": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "method": { "type": "string", "enum": ["BINANCE_PAY", "ONCHAIN"] },
                        "transactionId": { "type": "string" },
                        "txHash": { "type": "string" },
                        "amount": { "type": "number" },
                        "currency": { "type": "string" },
                        "network": { "type": "string" },
                        "transactionTime": { "type": "integer" }
                      }
                    },
                    "criteria": { "type": "object" }
                  }
                }
              }
            }
          },
          "400": { "description": "Credentials/input tidak valid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/debug/pay-history": {
      "post": {
        "summary": "Riwayat Binance Pay",
        "description": "Ambil riwayat transaksi Binance Pay (masuk + keluar). Mode cepat pakai 'hours' (max 168). Mode 'full' loop semua transaksi mundur sampai 'days' (max 365) dengan dedupe otomatis.",
        "tags": ["Debug"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "hours": { "type": "number", "example": 24, "description": "Jam ke belakang (mode cepat, max 168)" },
                  "full": { "type": "boolean", "example": false, "description": "Aktifkan mode full (loop semua)" },
                  "days": { "type": "number", "example": 90, "description": "Hari ke belakang untuk mode full (max 365)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Daftar transaksi Pay" },
          "400": { "description": "Credentials belum diisi" },
          "502": { "description": "Error dari Binance" }
        }
      }
    },
    "/debug/deposit-history": {
      "post": {
        "summary": "Riwayat deposit on-chain",
        "description": "Ambil riwayat deposit crypto masuk (on-chain) ke akun Binance.",
        "tags": ["Debug"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "hours": { "type": "number", "example": 24, "description": "Jam ke belakang (max 168)" },
                  "coin": { "type": "string", "example": "USDT", "description": "Filter per coin (opsional)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Daftar deposit" },
          "400": { "description": "Credentials belum diisi" },
          "502": { "description": "Error dari Binance" }
        }
      }
    },
    "/debug/withdraw-history": {
      "post": {
        "summary": "Riwayat withdraw on-chain",
        "description": "Ambil riwayat penarikan crypto keluar (on-chain) dari akun Binance.",
        "tags": ["Debug"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "hours": { "type": "number", "example": 24, "description": "Jam ke belakang (max 168)" },
                  "coin": { "type": "string", "example": "USDT", "description": "Filter per coin (opsional)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Daftar withdraw" },
          "400": { "description": "Credentials belum diisi" },
          "502": { "description": "Error dari Binance" }
        }
      }
    },
    "/debug/balances": {
      "post": {
        "summary": "Saldo spot",
        "description": "Ambil saldo semua aset spot yang lebih dari 0 (snapshot saat ini).",
        "tags": ["Debug"],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": { "description": "Daftar saldo aset" },
          "400": { "description": "Credentials belum diisi" },
          "502": { "description": "Error dari Binance" }
        }
      }
    },
    "/debug/overview": {
      "post": {
        "summary": "Overview akun lengkap",
        "description": "Ambil semua isi akun sekaligus: saldo + timeline gabungan (Pay, deposit, withdraw, convert, dividen) diurut dari terbaru. Semua sumber ditarik paralel; sumber yang gagal tidak menggagalkan yang lain.",
        "tags": ["Debug"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "days": { "type": "number", "example": 30, "description": "Hari ke belakang (max 90)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Saldo + timeline aktivitas + status tiap sumber" },
          "400": { "description": "Credentials belum diisi" }
        }
      }
    }
  }
}